+++ src/country.rs
@@ -14,5 +14,6 @@ pub enum Code {
     LU,
     PT,
     US,
-    DE
+    DE,
+    MX,
 }